Victor Doom source port
Runs id Software's Doom in the anim process (I didn't want to attempt
to run it as a separate program to avoid bricking the robot.)
This port is based on the SFML-DOOM source port, which runs Doom using
the SFML game library. I kept the same API as SFML to minimize changes
to the game source, but re-implemented SFML methods related to
rendering, playing audio, and taking input, so that ultimately, the anim
process could query the next image to be drawn or sound to be played.
To start DOOM, shake the robot and wait a moment. To play:
 + use the backpack button to fire
 + tilt the robot away from you to move forward
 + tilt the robot toward you to move backward
 + tilt the robot to the left/right to rotate left/right
 + raise the lift to open doors, etc.

NOTE: Both Doom and Timidity (Midi-to-wave library) are
      under GPL. DON'T SHIP! even as an awesome easter egg!